[#39] Base test structure implemented using unittest and pytest
