+++ .github/workflows/ci-tests.yml
@@ -1,0 +1,36 @@
+name: CI
+
+on:
+  pull_request:
+    branches: [ "dev" ]
+
+permissions:
+  contents: read
+
+jobs:
+  ci-test:
+
+    runs-on: ubuntu-latest
+
+    steps:
+    - uses: actions/checkout@v4
+    - name: Set up Python 3.12
+      uses: actions/setup-python@v3
+      with:
+        python-version: "3.12"
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install cython ase numpy==1.26.4 scipy matplotlib mpi4py ipython \
+        sphinx sphinx_rtd_theme numpy==1.26.4 mpmath asap3 pytest
+
+    - name: CI-tests
+      run: |
+        cd Tests
+        pytest --junit-xml=report.xml --color=yes
+        cd ..
+
+    - name: Upload-results
+      uses: actions/upload-artifact@v4
+      with:
+        path: Tests\*.xml
