+++ Outputs/.gitignore
@@ -1,0 +1,2 @@
+*.traj
+*.log
