--- POSCAR
@@ -1,10 +1,0 @@
-FCC Cu
-3.61
-0.5 0.5 0.0
-0.5 0.0 0.5
-0.0 0.5 0.5
-Cu
-1
-Direct
-0.00 0.00 0.00
-
