+++ Settings/settings.json
@@ -1,0 +1,19 @@
+{
+  "Ensemble": "NVE",
+
+  "Temperature": 2000,
+  "Timestep": 1,
+  "Number_of_steps": 2000,
+  "Interval": 50,
+  "Potential": "EMT",
+  "Attachments": ["energy"],
+
+  "Friction": 1,
+  "Pressure": 0,
+  "Compressibility": 1e-7,
+
+  "EquilSteps": 5000,
+  "Supercells": [3,3,3],
+  "Output_file" : "Outputs/output"
+
+}
