+++ Tests/TESTPOSCAR
@@ -1,0 +1,9 @@
+Fcc Copper
+3.57
+0.0 0.5 0.5
+0.5 0.0 0.5
+0.5 0.5 0.0
+Cu
+1
+Direct
+0.00 0.00 0.00
