+++ Tests/pytest.ini
@@ -1,0 +1,7 @@
+[pytest]
+log_cli = true
+log_cli_level = info
+log_format = %(asctime)s.%(msecs)03d %(levelname)s %(message)s
+log_date_format = %Y-%m-%d %H:%M:%S
+testpaths = Test*
+
