+++ Tests/testoutput.log
@@ -1,0 +1,41 @@
+Time[ps]    Etot/N[eV]   Epot/N[eV]   Ekin/N[eV]    T[K]
+0.0500           0.4847       0.2676       0.2171  1679.6
+0.1000           0.4847       0.2512       0.2335  1806.3
+0.1500           0.4847       0.2278       0.2568  1987.0
+0.2000           0.4847       0.2423       0.2424  1874.9
+0.2500           0.4847       0.2406       0.2441  1888.3
+0.3000           0.4847       0.2362       0.2484  1922.0
+0.3500           0.4847       0.2250       0.2597  2008.9
+0.4000           0.4847       0.2059       0.2788  2156.5
+0.4500           0.4847       0.2821       0.2025  1566.8
+0.5000           0.4847       0.2811       0.2036  1575.2
+0.5500           0.4847       0.2533       0.2314  1790.3
+0.6000           0.4847       0.2546       0.2301  1779.9
+0.6500           0.4847       0.2540       0.2307  1784.9
+0.7000           0.4847       0.2362       0.2485  1922.3
+0.7500           0.4846       0.2069       0.2778  2149.0
+0.8000           0.4847       0.2530       0.2317  1792.2
+0.8500           0.4847       0.1974       0.2873  2222.3
+0.9000           0.4847       0.2142       0.2704  2092.1
+0.9500           0.4847       0.2421       0.2425  1876.2
+1.0000           0.4847       0.2869       0.1978  1530.5
+1.0500           0.4847       0.2221       0.2626  2031.5
+1.1000           0.4847       0.2323       0.2524  1952.4
+1.1500           0.4847       0.2380       0.2466  1907.9
+1.2000           0.4847       0.2747       0.2100  1624.8
+1.2500           0.4846       0.1883       0.2963  2292.5
+1.3000           0.4847       0.2298       0.2549  1971.7
+1.3500           0.4847       0.2788       0.2059  1593.2
+1.4000           0.4847       0.2366       0.2481  1919.3
+1.4500           0.4847       0.2309       0.2537  1962.8
+1.5000           0.4847       0.2759       0.2088  1615.6
+1.5500           0.4847       0.2361       0.2486  1923.0
+1.6000           0.4847       0.2745       0.2101  1625.7
+1.6500           0.4847       0.2521       0.2325  1799.0
+1.7000           0.4846       0.2252       0.2594  2007.2
+1.7500           0.4847       0.1968       0.2879  2227.0
+1.8000           0.4847       0.2543       0.2304  1782.6
+1.8500           0.4847       0.2500       0.2347  1815.7
+1.9000           0.4847       0.2513       0.2334  1805.8
+1.9500           0.4847       0.2217       0.2630  2034.6
+2.0000           0.4847       0.2838       0.2009  1554.4
