--- requirements.txt
@@ -1,4 +1,0 @@
-numpy
-ase
-asap3
-matplotlib
